Module clojurust::clojure::rust::objects

Modules

Associative
Protocol Associative

Callable

Class
Class module

Collection

Comparable
Protocol Comparable

Comparator

Counted

FnMethodNative
Anonymous Native (Rust) method with a defined arity

Function
Anonymous Function with multi-arity

Globals
Stores for Rust Objects identified by unique string name

ILookup

IObject

Indexed

Iterable

List

Map
Protocol Map

Member
Access to Members getters and setters of Objects inner structure.

Number
All numbers and Number protocol

ObjResult
Standard Errors for the library

Object
Defines Rust dynamic Objects.

Protocol
Protocol is the eqivalent of traits.

Prototype
Prototype defines the methods of Protocols

RandomAccess

Reversible
Protocol Reversible

Runnable

Serializable

Set

String
Defines library’s dynamic Strings.

Thread

Unique
Unique association of names with unique id